Create Marketing Targets Project
The objective of the project was to predict if a client would subscribe to a long term deposit or not. The models used were:
Logistic Regression, Neural Networks, Forest with data transformations, Forest without data transformations and Decision trees.
The champion model was Forest without data transformation with. The model was chosen based on the KS (Youden) for the Validate partition (0.74). 90.55% of the Validate partition was correctly classified using the Forest model.